moved almost_equal into testutils header file
